restruct and fix CamelCase syntax in file names
On case sesitive systems there were a lot of compilation
errors "No such file or directory" due to wrong upper case
letter in file names.

As part of fix them some changes on the code style were made:

  - correct CamelCase in all file names
  - more or missing new lines for seperaition
  - remove trailing white space on line end
  - rename header from .h -> .hpp
  - resort order of inclusion (system header first)
  - global search for public interface header
  - remove #pragma once from main files
  - remove #pragma once from header files and add ifndef-
    define-endif rules instead

Related-to-issue: #4
