+++ Libusb++.vcxproj
@@ -90,22 +90,22 @@
     </Lib>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="headers\Configuration.h" />
-    <ClInclude Include="headers\Device.h" />
-    <ClInclude Include="headers\Endpoint.h" />
-    <ClInclude Include="headers\Endpointdefs.h" />
-    <ClInclude Include="headers\Interface.h" />
-    <ClInclude Include="headers\Libusbpp.h" />
-    <ClInclude Include="headers\Transfer.h" />
-    <ClInclude Include="headers\Transferdefs.h" />
-    <ClInclude Include="headers\UsbException.h" />
+    <ClInclude Include="headers\libusbpp\Configuration.h" />
+    <ClInclude Include="headers\libusbpp\Device.h" />
+    <ClInclude Include="headers\libusbpp\Endpoint.h" />
+    <ClInclude Include="headers\libusbpp\EndpointDefs.h" />
+    <ClInclude Include="headers\libusbpp\Exception.h" />
+    <ClInclude Include="headers\libusbpp\Interface.h" />
+    <ClInclude Include="headers\libusbpp\Transfer.h" />
+    <ClInclude Include="headers\libusbpp\TransferDefs.h" />
+    <ClInclude Include="headers\libusbpp.h" />
     <ClInclude Include="src\ConfigurationImpl.h" />
     <ClInclude Include="src\DeviceImpl.h" />
     <ClInclude Include="src\EndpointImpl.h" />
     <ClInclude Include="src\InterfaceImpl.h" />
     <ClInclude Include="src\LibusbImpl.h" />
     <ClInclude Include="src\TransferImpl.h" />
-    <ClInclude Include="src\wideconvert.h" />
+    <ClInclude Include="src\Wideconvert.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\Configuration.cpp" />
@@ -114,14 +114,14 @@
     <ClCompile Include="src\DeviceImpl.cpp" />
     <ClCompile Include="src\Endpoint.cpp" />
     <ClCompile Include="src\EndpointImpl.cpp" />
+    <ClCompile Include="src\Exception.cpp" />
     <ClCompile Include="src\Interface.cpp" />
     <ClCompile Include="src\InterfaceImpl.cpp" />
     <ClCompile Include="src\LibusbImpl.cpp" />
-    <ClCompile Include="src\Libusbpp.cpp" />
+    <ClCompile Include="src\libusbpp.cpp" />
     <ClCompile Include="src\Transfer.cpp" />
     <ClCompile Include="src\TransferImpl.cpp" />
-    <ClCompile Include="src\UsbException.cpp" />
-    <ClCompile Include="src\wideconvert.cpp" />
+    <ClCompile Include="src\Wideconvert.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
