+++ headers/libusbpp/Configuration.hpp
@@ -1,8 +1,32 @@
-#pragma once
+/*
+ * Copyright (C) 2012, Anthony Clay, ZarthCode LLC, all rights reserved.
+ * Copyright (C) 2016, Stephan Linz, Li-Pro.Net, all rights reserved.
+ *
+ * This file is part of the LibUSB C++ wrapper library (libusbpp).
+ *
+ * libusbpp is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * libusbpp is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with libusbpp.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef LIBUSBPP_CONFIGURATION_HPP
+#define LIBUSBPP_CONFIGURATION_HPP
+
 #include <memory>
 #include <string>
 #include <stdint.h>
-#include "interface.h"
+
+#include <libusbpp/Interface.hpp>
+
 
 namespace LibUSB
 {
@@ -68,3 +92,5 @@ namespace LibUSB
 	};
 
 }
+
+#endif // LIBUSBPP_CONFIGURATION_HPP
