+++ headers/libusbpp/Device.hpp
@@ -1,9 +1,32 @@
-#pragma once
-#include "Configuration.h"
-#include "Endpoint.h"
+/*
+ * Copyright (C) 2012, Anthony Clay, ZarthCode LLC, all rights reserved.
+ * Copyright (C) 2016, Stephan Linz, Li-Pro.Net, all rights reserved.
+ *
+ * This file is part of the LibUSB C++ wrapper library (libusbpp).
+ *
+ * libusbpp is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * libusbpp is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with libusbpp.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef LIBUSBPP_DEVICE_HPP
+#define LIBUSBPP_DEVICE_HPP
+
 #include <stdint.h>
 #include <string>
 #include <memory>
+
+#include <libusbpp/Configuration.hpp>
+#include <libusbpp/Endpoint.hpp>
 
 
 namespace LibUSB
@@ -30,7 +53,7 @@ namespace LibUSB
 
 		/// Opens the device.
 		void Open();
-	
+
 	// Device Descriptor
 
 		/// USB specification release number
@@ -39,10 +62,10 @@ namespace LibUSB
 		/// USB-IF class code
 		uint8_t DeviceClass();
 
-		/// USB-IF subclass code for the device, qualified by the bDeviceClass value. 
+		/// USB-IF subclass code for the device, qualified by the bDeviceClass value.
 		uint8_t DeviceSubclass();
 
-		/// USB-IF protocol code for the device, qualified by the bDeviceClass and bDeviceSubClass values. 
+		/// USB-IF protocol code for the device, qualified by the bDeviceClass and bDeviceSubClass values.
 		uint8_t DeviceProtocol();
 
 		/// USB-IF vendor ID
@@ -82,7 +105,7 @@ namespace LibUSB
 
 		/// Transfers need access to the transfer event notification method of their target device.
 		friend class TransferImpl;
-		
+
 		/// Notification of a completed transfer
 		/// \warning This function can be called from other threads when using asynchronous transfers!
 		virtual void TransferEventNotification(std::shared_ptr<Transfer> pCompletedTransfer);
@@ -96,8 +119,10 @@ namespace LibUSB
 		/// Device Implementation Object
 		std::shared_ptr<DeviceImpl> m_pDeviceImpl_;
 
-		
-		
+
+
 	};
 
 }
+
+#endif // LIBUSBPP_DEVICE_HPP
