+++ headers/libusbpp/Endpoint.hpp
@@ -1,10 +1,31 @@
-#pragma once
+/*
+ * Copyright (C) 2012, Anthony Clay, ZarthCode LLC, all rights reserved.
+ * Copyright (C) 2016, Stephan Linz, Li-Pro.Net, all rights reserved.
+ *
+ * This file is part of the LibUSB C++ wrapper library (libusbpp).
+ *
+ * libusbpp is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * libusbpp is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with libusbpp.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
-#include "Endpointdefs.h"
+#ifndef LIBUSBPP_ENDPOINT_HPP
+#define LIBUSBPP_ENDPOINT_HPP
 
 #include <memory>
 
-#include "Transfer.h"
+#include <libusbpp/Transfer.hpp>
+#include <libusbpp/EndpointDefs.hpp>
+
 
 namespace LibUSB
 {
@@ -49,7 +70,7 @@ namespace LibUSB
 		/// Creates a new data transfer object
 		std::shared_ptr<Transfer> CreateTransfer();
 
-		
+
 
 	protected:
 	private:
@@ -59,3 +80,5 @@ namespace LibUSB
 	};
 
 }
+
+#endif // LIBUSBPP_ENDPOINT_HPP
