+++ src/ConfigurationImpl.cpp
@@ -1,14 +1,35 @@
-#include "ConfigurationImpl.h"
-#include "deviceimpl.h"
-#include "Interface.h"
-#include "Interfaceimpl.h"
-#include "usbexception.h"
-
+/*
+ * Copyright (C) 2012, Anthony Clay, ZarthCode LLC, all rights reserved.
+ * Copyright (C) 2016, Stephan Linz, Li-Pro.Net, all rights reserved.
+ *
+ * This file is part of the LibUSB C++ wrapper library (libusbpp).
+ *
+ * libusbpp is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * libusbpp is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with libusbpp.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
 #include <stdexcept>
 #include <algorithm>
 #include <xlocale>
-#include "wideconvert.h"
+
+#include <libusbpp/Interface.hpp>
+#include <libusbpp/Exception.hpp>
+
+#include "ConfigurationImpl.hpp"
+#include "DeviceImpl.hpp"
+#include "InterfaceImpl.hpp"
+#include "Wideconvert.hpp"
+
 
 LibUSB::ConfigurationImpl::ConfigurationImpl( libusb_config_descriptor* pConfigDescriptor, std::weak_ptr<DeviceImpl> pDeviceImpl)
 {
@@ -51,7 +72,7 @@ std::wstring LibUSB::ConfigurationImpl::DescriptorString( void ) const
 
 
 	std::wstring resultStr;
-	
+
 	try
 	{
 		resultStr = m_pDeviceImpl.lock()->getStringDescriptorW(m_pConfigDescriptor->iConfiguration);
@@ -77,7 +98,7 @@ uint8_t LibUSB::ConfigurationImpl::getMaxPower() const
 {
 
 	return m_pConfigDescriptor->MaxPower;
-	
+
 }
 
 bool LibUSB::ConfigurationImpl::isSelfPowered() const
@@ -130,7 +151,7 @@ int LibUSB::ConfigurationImpl::NumInterfaces() const
 std::shared_ptr<LibUSB::Interface> LibUSB::ConfigurationImpl::getInterfaceByIndex( int index ) const
 {
 	/// \note I'm at a loss determining if the (array) index is required to correspond to the interface number.
-	
+
 	// Locate the indicated interface
 	if (index >= m_pConfigDescriptor->bNumInterfaces)
 	{
@@ -144,7 +165,7 @@ std::shared_ptr<LibUSB::Interface> LibUSB::ConfigurationImpl::getInterfaceByInde
 	// Create the InterfaceImpl object
 	const libusb_interface *pInterface = &(m_pConfigDescriptor->interface[index]);
 	std::shared_ptr<InterfaceImpl> pInterfaceImpl = std::make_shared<InterfaceImpl>(pInterface, m_pDeviceImpl);
-	
+
 	// Create the interface object.
 	return std::make_shared<Interface>(pInterfaceImpl);
 
@@ -170,7 +191,7 @@ std::shared_ptr<LibUSB::Interface> LibUSB::ConfigurationImpl::getInterface( int 
 {
 
 	// From what I can tell, interface number 0 isn't exactly invalid/disallowed - so no range check.
-	
+
 	InterfaceMap_t::const_iterator itInterface = m_InterfaceMap.find(InterfaceNumber);
 
 	if (itInterface == m_InterfaceMap.end())
@@ -179,6 +200,6 @@ std::shared_ptr<LibUSB::Interface> LibUSB::ConfigurationImpl::getInterface( int 
 	}
 
 	return itInterface->second;
-	
+
 }
