+++ src/ConfigurationImpl.hpp
@@ -1,9 +1,33 @@
-#pragma once
+/*
+ * Copyright (C) 2012, Anthony Clay, ZarthCode LLC, all rights reserved.
+ * Copyright (C) 2016, Stephan Linz, Li-Pro.Net, all rights reserved.
+ *
+ * This file is part of the LibUSB C++ wrapper library (libusbpp).
+ *
+ * libusbpp is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * libusbpp is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with libusbpp.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef __LIBUSBPP_CONFIGURATION_IMPL_HPP
+#define __LIBUSBPP_CONFIGURATION_IMPL_HPP
+
 #include <memory>
-#include <libusb/libusb.h>
 #include <string>
-#include "interface.h"
 #include <map>
+
+#include <libusb/libusb.h>
+
+#include <libusbpp/Interface.hpp>
 
 
 namespace LibUSB
@@ -23,8 +47,8 @@ namespace LibUSB
 	class ConfigurationImpl
 	{
 	public:
-		
-		/// Constructor - initializes 
+
+		/// Constructor - initializes
 		ConfigurationImpl(libusb_config_descriptor* pConfigDescriptor, std::weak_ptr<DeviceImpl> pParentDeviceImpl);
 		~ConfigurationImpl();
 
@@ -83,3 +107,5 @@ namespace LibUSB
 	};
 
 }
+
+#endif // __LIBUSBPP_CONFIGURATION_IMPL_HPP
