+++ src/DeviceImpl.hpp
@@ -1,10 +1,34 @@
-#pragma once
+/*
+ * Copyright (C) 2012, Anthony Clay, ZarthCode LLC, all rights reserved.
+ * Copyright (C) 2016, Stephan Linz, Li-Pro.Net, all rights reserved.
+ *
+ * This file is part of the LibUSB C++ wrapper library (libusbpp).
+ *
+ * libusbpp is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * libusbpp is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with libusbpp.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
-#include "Device.h"
-#include <libusb/libusb.h>
+#ifndef __LIBUSBPP_DEVICE_IMPL_HPP
+#define __LIBUSBPP_DEVICE_IMPL_HPP
+
 #include <memory>
 #include <string>
 #include <map>
+
+#include <libusb/libusb.h>
+
+#include <libusbpp/Device.hpp>
+
 
 namespace LibUSB
 {
@@ -76,7 +100,7 @@ namespace LibUSB
 
 	private:
 
-	
+
 		/// Weak_ptr collection of other configuration objects.
 		std::map<uint8_t, std::weak_ptr<Configuration>> m_ConfigurationMap;
 
@@ -99,3 +123,5 @@ namespace LibUSB
 
 
 }
+
+#endif // __LIBUSBPP_DEVICE_IMPL_HPP
