+++ src/EndpointImpl.hpp
@@ -1,11 +1,34 @@
-#pragma once
+/*
+ * Copyright (C) 2012, Anthony Clay, ZarthCode LLC, all rights reserved.
+ * Copyright (C) 2016, Stephan Linz, Li-Pro.Net, all rights reserved.
+ *
+ * This file is part of the LibUSB C++ wrapper library (libusbpp).
+ *
+ * libusbpp is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * libusbpp is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with libusbpp.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
-#include <libusb/libusb.h>
+#ifndef __LIBUSBPP_ENDPOINT_IMPL_HPP
+#define __LIBUSBPP_ENDPOINT_IMPL_HPP
+
 #include <memory>
 #include <stdint.h>
 
-#include "Endpointdefs.h"
-#include "Transfer.h"
+#include <libusb/libusb.h>
+
+#include <libusbpp/EndpointDefs.hpp>
+#include <libusbpp/Transfer.hpp>
+
 
 namespace LibUSB
 {
@@ -68,3 +91,5 @@ namespace LibUSB
 	};
 
 }
+
+#endif // __LIBUSBPP_ENDPOINT_IMPL_HPP
