+++ src/Interface.cpp
@@ -1,15 +1,34 @@
-#pragma once
+/*
+ * Copyright (C) 2012, Anthony Clay, ZarthCode LLC, all rights reserved.
+ * Copyright (C) 2016, Stephan Linz, Li-Pro.Net, all rights reserved.
+ *
+ * This file is part of the LibUSB C++ wrapper library (libusbpp).
+ *
+ * libusbpp is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * libusbpp is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with libusbpp.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
-#include "Interface.h"
-#include "Interfaceimpl.h"
+#include <libusb.h>
 
-#include <libusb/libusb.h>
+#include <libusbpp/Interface.hpp>
+
+#include "InterfaceImpl.hpp"
 
 
 LibUSB::Interface::Interface( std::shared_ptr<InterfaceImpl> pImpl )
 	:m_pInterfaceImpl(pImpl)
 {
-	
+
 }
