+++ src/InterfaceImpl.hpp
@@ -1,10 +1,33 @@
-#pragma once
-#include <libusb/libusb.h>
+/*
+ * Copyright (C) 2012, Anthony Clay, ZarthCode LLC, all rights reserved.
+ * Copyright (C) 2016, Stephan Linz, Li-Pro.Net, all rights reserved.
+ *
+ * This file is part of the LibUSB C++ wrapper library (libusbpp).
+ *
+ * libusbpp is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * libusbpp is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with libusbpp.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef __LIBUSBPP_INTERFACE_IMPL_HPP
+#define __LIBUSBPP_INTERFACE_IMPL_HPP
+
 #include <string>
 #include <memory>
 #include <map>
-#include "Endpoint.h"
 
+#include <libusb/libusb.h>
+
+#include <libusbpp/Endpoint.hpp>
 
 
 namespace LibUSB
@@ -48,7 +71,7 @@ namespace LibUSB
 
 		/// Returns TRUE if the interface is claimed
 		bool isClaimed()const;
-		
+
 		/*!
 		 * \brief Selects an alternate interface setting.
 		 * 
@@ -98,3 +121,5 @@ namespace LibUSB
 	};
 
 }
+
+#endif // __LIBUSBPP_INTERFACE_IMPL_HPP
