+++ src/Transfer.cpp
@@ -1,5 +1,26 @@
-#include "Transfer.h"
-#include "TransferImpl.h"
+/*
+ * Copyright (C) 2012, Anthony Clay, ZarthCode LLC, all rights reserved.
+ * Copyright (C) 2016, Stephan Linz, Li-Pro.Net, all rights reserved.
+ *
+ * This file is part of the LibUSB C++ wrapper library (libusbpp).
+ *
+ * libusbpp is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * libusbpp is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with libusbpp.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include <libusbpp/Transfer.hpp>
+
+#include "TransferImpl.hpp"
 
 
 LibUSB::Transfer::Transfer( std::shared_ptr<LibUSB::TransferImpl> pTransferImpl )
@@ -36,7 +57,7 @@ void LibUSB::Transfer::SetTimeout( std::chrono::milliseconds timeout )
 	}
 
 	m_pTransferImpl->setTimeout(timeout);
-	
+
 }
 
 void LibUSB::Transfer::setTransferBuffer( std::shared_ptr<unsigned char> pBuffer, size_t bufferSize )
@@ -103,7 +124,7 @@ void LibUSB::Transfer::Start()
 	catch (...)
 	{
 		throw;
-	
+
 	}
 
 }
@@ -120,7 +141,7 @@ bool LibUSB::Transfer::isComplete()
 
 	if (m_AsynchronousTransferPending)
 	{
-		
+
 		// Check the status of the thread.
 		if (*m_TransferThreadRunning)
 		{
@@ -134,10 +155,10 @@ bool LibUSB::Transfer::isComplete()
 
 	}
 
-	
+
 
 	return m_pTransferImpl->isComplete();
-	
+
 }
 
 size_t LibUSB::Transfer::BytesTransferred() const
@@ -152,7 +173,7 @@ size_t LibUSB::Transfer::BytesTransferred() const
 	}
 
 	return m_pTransferImpl->bytesTransferred();
-	
+
 }
 
 void LibUSB::Transfer::Reset()
@@ -204,10 +225,10 @@ void LibUSB::Transfer::AsyncStart()
 	m_TransferFuture = TransferPromise->get_future();
 
 	auto pPromisePending = std::make_shared<bool>(true);
-	
+
 	// Share a boolean as a signal.
 	m_TransferThreadRunning = pPromisePending;
-	
+
 	std::shared_ptr<LibUSB::TransferImpl> pTransferImplementation = m_pTransferImpl;
 
 	std::thread transferThread([=]()
@@ -222,16 +243,16 @@ void LibUSB::Transfer::AsyncStart()
 					catch (...)
 					{
 						TransferPromise->set_exception(std::current_exception());
-					
+
 					}
 
 					*pPromisePending = false;
-					
+
 				});
 
 	// Let it complete.
 	transferThread.detach();
-	
+
 }
 
 bool LibUSB::Transfer::WaitForCompletion()
@@ -280,14 +301,14 @@ void LibUSB::ControlTransfer::SetupPacket( uint8_t Request, uint16_t wValue, uin
 
 	std::static_pointer_cast<ControlTransferImpl>(m_pTransferImpl)->SetupPacket(Request, wValue, wIndex, transferDirection, requestType, recipient);
 
-	
+
 }
 
 LibUSB::ControlTransfer::ControlTransfer( std::shared_ptr<TransferImpl> pTransferImpl )
 	: Transfer(pTransferImpl)
 {
 
-	
+
 }
 
 LibUSB::ControlTransfer::~ControlTransfer()
@@ -299,7 +320,7 @@ LibUSB::InterruptTransfer::InterruptTransfer( std::shared_ptr<TransferImpl> pTra
 	: Transfer(pTransferImpl)
 {
 
-	
+
 }
 
 LibUSB::InterruptTransfer::~InterruptTransfer()
@@ -326,13 +347,13 @@ LibUSB::IsochronousTransfer::IsochronousTransfer( std::shared_ptr<LibUSB::Transf
 	:LibUSB::Transfer(pTransferImpl)
 {
 
-	
+
 }
 
 LibUSB::IsochronousTransfer::~IsochronousTransfer()
 {
 
-	
+
 }
 
 void LibUSB::IsochronousTransfer::setNumPackets( int Packets )
