--- src/wideconvert.h
@@ -1,15 +1,0 @@
-#include <string>
-
-
-namespace LibUSB
-{
-
-	namespace Util
-	{
-	// Prototype for conversion functions
-		std::wstring StringToWString(const std::string& s);
-		std::string WStringToString(const std::wstring& s);
-		
-		
-	}
-}
